refactor: update components files according to name convention
