+++ src/lib/components/card/card-sketch.svelte
