+++ src/lib/components/input-number/input-number.svelte
