+++ src/lib/components/input-range/index.ts
@@ -1,1 +1,1 @@
-export { default as Range } from "./Range.svelte";
+export { default as Range } from "./input-range.svelte";
