+++ src/lib/components/input-range/input-range.module.css
