+++ src/lib/components/menu-indicator/menu-indicator.svelte
