+++ src/lib/components/radio/index.ts
@@ -1,2 +1,2 @@
-export { default as Group } from "./RadioGroup.svelte";
-export { default as Button } from "./RadioButton.svelte";
+export { default as Group } from "./radio.group.svelte";
+export { default as Button } from "./radio.button.svelte";
